backend/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose
pydantic
fpdf
python-multipart
